refactor(langgraph-python): rename A2UI files and clean up suggestions
- Delete generate_form / Event Registration (Declarative UI)
- Rename a2ui_fixed → a2ui_fixed_schema
- Rename a2ui_streaming → a2ui_fixed_schema_streaming
- Rename a2ui_dynamic → a2ui_dynamic_schema
- Reorder suggestions: streaming before fixed schema
- Update imports in main.py
