+++ examples/integrations/langgraph-python/apps/agent/main.py
@@ -9,14 +9,13 @@ from langchain_openai import ChatOpenAI
 
 from src.query import query_data
 from src.todos import AgentState, todo_tools
-from src.form import generate_form
-from src.a2ui_fixed import search_flights
-from src.a2ui_streaming import search_flights_streaming
-from src.a2ui_dynamic import generate_a2ui
+from src.a2ui_fixed_schema import search_flights
+from src.a2ui_fixed_schema_streaming import search_flights_streaming
+from src.a2ui_dynamic_schema import generate_a2ui
 
 agent = create_agent(
     model="openai:gpt-4.1",
-    tools=[query_data, *todo_tools, generate_form, search_flights, search_flights_streaming, generate_a2ui],
+    tools=[query_data, *todo_tools, search_flights, search_flights_streaming, generate_a2ui],
     middleware=[CopilotKitMiddleware()],
     state_schema=AgentState,
     system_prompt="""
