+++ examples/integrations/langgraph-python/apps/agent/src/a2ui_dynamic_schema.py
